09.18 ordersController /read 하는중

+++ src/main/java/com/example/demo/order/entity/Orders.java
@@ -42,5 +42,5 @@ public class Orders extends BaseEntity {
 	private String shipAddress;
 	
 	@Column(nullable = false)
-	private double totalPrice;
+	private int totalPrice;
 }
